+++ layout/vg/tess/SLTessPriorityq.c
@@ -208,7 +208,7 @@ PQhandle pqHeapInsert( TESSalloc* alloc, PriorityQHeap *pq, PQkey keyNew )
 
 		pq->max <<= 1;
 
-		pq->nodes = (PQnode *)alloc->memalloc( alloc->userData, (size_t) ((pq->max + 1) * sizeof( pq->nodes[0] )));
+		pq->nodes = (PQnode *)alloc->memalloc( alloc->userData, (unsigned int) ((pq->max + 1) * sizeof( pq->nodes[0] )));
 		if (pq->nodes == NULL) {
 			pq->nodes = saveNodes;	// restore ptr to free upon return
 			return INV_HANDLE;
@@ -216,7 +216,7 @@ PQhandle pqHeapInsert( TESSalloc* alloc, PriorityQHeap *pq, PQkey keyNew )
 			memcpy(pq->nodes, saveNodes, (size_t)((tmpSize + 1) * sizeof(pq->nodes[0])) );
 		}
 
-		pq->handles = (PQhandleElem *)alloc->memalloc( alloc->userData, (size_t) ((pq->max + 1) * sizeof( pq->handles[0] )));
+		pq->handles = (PQhandleElem *)alloc->memalloc( alloc->userData, (unsigned int) ((pq->max + 1) * sizeof( pq->handles[0] )));
 		if (pq->handles == NULL) {
 			pq->handles = saveHandles; // restore ptr to free upon return
 			return INV_HANDLE;
@@ -353,7 +353,7 @@ int pqInit( TESSalloc* alloc, PriorityQ *pq )
 	(pq->size * sizeof(pq->order[0])) );
 	*/
 	pq->order = (PQkey **)alloc->memalloc( alloc->userData,
-										  (size_t)((pq->size+1) * sizeof(pq->order[0])) );
+										  (unsigned int)((pq->size+1) * sizeof(pq->order[0])) );
 	/* the previous line is a patch to compensate for the fact that IBM */
 	/* machines return a null on a malloc of zero bytes (unlike SGI),   */
 	/* so we have to put in this defense to guard against a memory      */
@@ -434,7 +434,7 @@ PQhandle pqInsert( TESSalloc* alloc, PriorityQ *pq, PQkey keyNew )
 		// If the heap overflows, double its size.
 		int tmpSize = pq->max;
 		pq->max <<= 1;
-		pq->keys = (PQkey *)alloc->memalloc( alloc->userData, (size_t)(pq->max * sizeof( pq->keys[0] )));
+		pq->keys = (PQkey *)alloc->memalloc( alloc->userData, (unsigned int)(pq->max * sizeof( pq->keys[0] )));
 		if (pq->keys == NULL) {
 			pq->keys = saveKey;  // restore ptr to free upon return
 			return INV_HANDLE;
